服务器相关修改

+++ game_server/logic_server/src/main/java/db/cache/MulSearch.java
@@ -88,7 +88,7 @@ public class MulSearch<E> {
         int pos = 0;
 
         // 1. find
-        pos = (key.intValue() % hash_size);
+        pos = (int)(key.longValue() % hash_size);
         iterator = ports[pos].root;
         if (null == iterator) {
             return false;
